removed the old level js file
