+++ src/js/level1.js
@@ -1,175 +1,164 @@
-import Candy, {Colors, Shapes, Patterns } from './candy.js';
+import PathManager from "./SceneClasses/PathManager.js";
+import AnimationExecutor from "./SceneClasses/AnimationExecutor.js";
+import CommandManager from "./SceneClasses/CommandManager.js";
+import QueueManager from "./SceneClasses/QueueManager.js";
+import LevelHelper from "./SceneClasses/LevelHelper.js";
+
 export default class Level1 extends Phaser.Scene {
+  graphics;
+  pathManager;
+  animationExecutor;
+  commandManager;
+  currentLevel = "Level1";
 
-    graphics;
-    path1;
-    path2;
-    path3;
-    follower;
-    isMoving = false;
+  preload() {
+    this.load.image("background", "assets/background.png");
+    console.log(`[${this.currentLevel}] Preloading background image.`);
+  }
 
-    preload() {
-        // Load the background image
-        //TO-DO: Add Texture manager: https://docs.phaser.io/phaser/concepts/textures
-        //Example candy implementation 
-        const blueStripedCircle = new Candy(Colors.BLUE, Shapes.CIRCLE, Patterns.STRIPED, '../assets/candy_photos/blue_circle_striped.png');
-        //console.log(blueStripedCircle.imagePath === '../assets/blue_circle_striped.png');
-        this.load.image('background', 'assets/background.png');
-        this.load.image('follower', blueStripedCircle.imagePath); // Load the candy image
-        //this.load.image('follower', 'assets/follower.png'); // Optional: Load a follower sprite
-    }
+  initializeEditorWindow() {
+    LevelHelper.initializeEditorWindow(
+      this,
+      "moveDown\nmoveLeft\nmoveLeft\ndumpCandy\nmoveDown\nmoveRight\nmoveRight\ndumpCandy\nmoveDown\nmoveDown\ndumpCandy",
+    );
+  }
 
-    create() {
-        // Initialize the editor window
-        C4C.Editor.Window.init(this);   // Scene is passed in to this init function!
-        C4C.Editor.Window.open();
-        console.log("Text editor initialized.");
+  initializeBackgroundGraphics() {
+    this.add.image(400, 300, "background");
+    console.log(`[${this.currentLevel}] Background image added.`);
 
-        // Define interpreter commands
-        C4C.Interpreter.define("moveleft", () => {
-            //console.log("moveleft in text editor");
-            this.moveLeft();
-        });
+    this.graphics = this.add.graphics();
+    console.log(`[${this.currentLevel}] Graphics object created.`);
+  }
 
-        C4C.Interpreter.define("moveright", () => {
-            this.moveRight();
-        });
+  createLinesForConveyerBelt() {
+    this.pathManager.addLine("center", { x: 400, y: 100 }, { x: 400, y: 400 });
+    this.pathManager.addLineFrom("center", "left", { x: 200, y: 400 });
+    this.pathManager.addLineFrom('center', 'right', { x: 600, y: 400 });
+    this.pathManager.addLineFrom('center', 'leftDown', { x: 400, y: 500 });
+    // this.pathManager.addLineFrom('center', 'rightDown', { x: 600, y: 150 });
+  }
 
-        document.getElementById("enableCommands").addEventListener("click", (event) => {
-            let programText = C4C.Editor.getText();
-            console.log("Program text: ", programText);
-            C4C.Interpreter.run(programText);
-            runner.setProgram(programText);
-        });
+  createIncrementalCommands() {
+    LevelHelper.createIncrementalCommands(this.pathManager, {
+      moveLeft: (currentPos) => ({ x: currentPos.x - 100, y: currentPos.y }),
+      moveRight: (currentPos) => ({ x: currentPos.x + 100, y: currentPos.y }),
+      moveUp: (currentPos) => ({ x: currentPos.x, y: currentPos.y - 100 }),
+      moveDown: (currentPos) => ({ x: currentPos.x, y: currentPos.y + 100 }),
+    });
+  }
 
-        // Add the background image
-        this.add.image(400, 300, 'background'); // Center the background
+  setupLevelCandies() {
+    //Define the candies for this level
+    //TODO: Adjust this to use the Candy class!
+    const candies = [
+      { type: "blue-circle", id: 1 },
+      { type: "red-square", id: 2 },
+      { type: "green-triangle", id: 3 },
+    ];
 
-        this.graphics = this.add.graphics();
-        this.initializePaths();
-        this.initializeFollower();
-
-        // Add event listener to the button
-        document.getElementById("enableCommands").addEventListener("click", this.startTween);
-    }
-
-    initializePaths() {
-        // Create the path using 3 separate lines
-        const startline = new Phaser.Curves.Line([400, 0, 400, 300]);
-        const leftline = new Phaser.Curves.Line([400, 300, 300, 500]);
-        const rightline = new Phaser.Curves.Line([400, 300, 500, 500]);
-
-        this.path1 = this.add.path();
-        this.path1.add(startline);
-
-        this.path2 = this.add.path();
-        this.path2.add(leftline);
-
-        this.path3 = this.add.path();
-        this.path3.add(rightline);
-    }
-
-    initializeFollower() {
-        this.follower = { t: 0, vec: new Phaser.Math.Vector2() };
-    }
-
-    startTween = () => {
-        this.follower.t = 0;
-        this.isMoving = true;
-        console.log("isMoving: ", this.isMoving);
-        this.tweens.add({
-            targets: this.follower,
-            t: 1,
-            ease: 'Linear',
-            duration: 1000,
-            onUpdate: () => {
-                this.path1.getPoint(this.follower.t, this.follower.vec);
-            },
-            onComplete: () => {
-                console.log("Start Path complete!");
-                this.handlePathCompletion();
-            }
-        });
+    //Define goal positions for each candy type. Again, adjust to using the Candy class
+    const goalPositions = {
+      "blue-circle": { x: 200, y: 400 }, // Left bin
+      "red-square": { x: 600, y: 400 }, // Right bin
+      "green-triangle": { x: 400, y: 500 }, // Bottom bin
     };
 
-    handlePathCompletion() {
-        // //Logic to determine the next move based on user input or pseudo code
-        // const nextMove = this.getNextMove(); // Implement this function to determine the next move
-        // if (nextMove === "left") {
-        //     this.moveLeft();
-        // } else if (nextMove === "right") {
-        //     this.moveRight();
-        // } else {
-        //     this.isMoving = false;
-        // }
-    }
+    // Set up callbacks for candy completion
+    this.pathManager.setCallbacks(
+      (candy) => this.onCandySuccess(candy),
+      (candy, position) => this.onCandyFailed(candy, position),
+    );
 
-    // getNextMove() {
-    // }
+    this.pathManager.setupCandyQueueAndGoalPositions(candies, goalPositions);
+  }
 
-    moveLeft = () => {
-        console.log("Move left function called");
-        this.follower.t = 1;
-        this.tweens.add({
-            targets: this.follower,
-            t: 2,
-            ease: 'Linear',
-            duration: 1000,
-            onUpdate: () => {
-                this.path2.getPoint(this.follower.t - 1, this.follower.vec);
-            },
-            onComplete: () => {
-                console.log("Left Path complete!");
-                this.handlePathCompletion();
-            }
-        });
-    };
+  //Having these two methods below in Level1.js is fine for now- to be discussed if we just
+  //want the same behavior for each case anyways, if so we can just export it to CommandManager
+  onCandySuccess(candy) {
+    LevelHelper.onCandySuccess(this, candy);
+  }
 
-    moveRight = () => {
-        console.log("Move right function called");
-        this.follower.t = 2.001;
-        this.tweens.add({
-            targets: this.follower,
-            t: 3,
-            ease: 'Linear',
-            duration: 1000,
-            onUpdate: () => {
-                this.path3.getPoint(this.follower.t - 2, this.follower.vec);
-            },
-            onComplete: () => {
-                console.log("Right Path complete!");
-                this.handlePathCompletion();
-            }
-        });
-    };
+  onCandyFailed(candy, position) {
+    LevelHelper.onCandyFailed(this, candy, position);
+  }
 
-    update() {
-        //Clear the graphics object
-        this.graphics.clear();
-        this.graphics.lineStyle(2, 0xffffff, 1);
+  defineInterpreterCommands() {
+    LevelHelper.defineInterpreterCommands(this.commandManager, {
+      immediate: {
+        sampleCommand: () => {
+          console.log(
+            "This is an example custom command, should run immediately",
+          );
+        },
+      },
+      queued: {
+        queuedCommand: () => {
+          console.log(
+            "This is an example custom command that is queued according to animation, should run in animation sequence",
+          );
+        },
+      },
+    });
+  }
 
-        //Draw the paths
-        this.path1.draw(this.graphics);
-        this.path2.draw(this.graphics);
-        this.path3.draw(this.graphics);
+  initializeRunCodeButton() {
+    LevelHelper.initializeRunCodeButton(
+      this,
+      this.setupLevelCandies.bind(this),
+      this.animationExecutor,
+      this.queueManager,
+    );
+  }
+  // Add a button to reset the level completely
+  initializeResetButton() {
+    LevelHelper.initializeResetButton(
+      this,
+      this.setupLevelCandies.bind(this),
+      this.animationExecutor,
+      this.queueManager,
+    );
+  }
 
-        //Get the position of the follower on the path
-        if (this.isMoving) {
-            if (this.follower.t <= 1) {
-                this.path1.getPoint(this.follower.t, this.follower.vec);
-            } else if (this.follower.t > 1 && this.follower.t <= 2) {
-                this.path2.getPoint(this.follower.t - 1, this.follower.vec);
-            } else if (this.follower.t > 2 && this.follower.t <= 3) {
-                this.path3.getPoint(this.follower.t - 2, this.follower.vec);
-            }
-        }
+  resetLevel() {
+    LevelHelper.resetLevel(
+      this,
+      this.setupLevelCandies.bind(this),
+      this.animationExecutor,
+      this.queueManager,
+    );
+  }
 
-        //Draw the follower as a red square
-        this.graphics.fillStyle(0xff0000, 1);
-        this.graphics.fillRect(this.follower.vec.x - 8, this.follower.vec.y - 8, 16, 16);
-    }
+  create() {
+    this.initializeEditorWindow();
+    this.initializeBackgroundGraphics();
+    this.pathManager = new PathManager(this);
+    this.animationExecutor = new AnimationExecutor(this, this.pathManager);
+    this.queueManager = new QueueManager(
+      this.pathManager,
+      this.animationExecutor,
+    );
+    this.commandManager = new CommandManager(
+      this,
+      this.pathManager,
+      this.animationExecutor,
+      this.queueManager,
+    );
+
+    //Set up the level
+    this.createLinesForConveyerBelt();
+    this.createIncrementalCommands();
+    this.setupLevelCandies();
+    this.defineInterpreterCommands();
+    this.initializeRunCodeButton();
+    this.initializeResetButton();
+  }
+
+  update() {
+    this.graphics.clear();
+    this.graphics.lineStyle(4, 0xffffff, 1);
+
+    this.pathManager.drawAll(this.graphics);
+    this.animationExecutor.drawFollower(this.graphics);
+  }
 }
-
-//For debugging for casey later...
-// const canvas = document.getElementById('my-custom-canvas');
-// if (canvas) {console.log("Found?");} else { console.log("Not found?"); }
